WIP change examples project structure and create mvn prj.
