+++ examples/producer/src/main/java/org/oracle/okafka/examples/Producer.java
@@ -5,7 +5,7 @@
 ** Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 */
 
-package producer;
+package org.oracle.okafka.examples;
 
 import java.util.Properties;
