+++ examples/producer/src/test/java/org/oracle/okafka/examples/ProducerTest.java
@@ -1,0 +1,20 @@
+package org.oracle.okafka.examples;
+
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class ProducerTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
+}
